Splited OrderAPI to its own package
