+++ backend/src/main/java/se/chalmers/dat265/group1/api/ApiController.java
@@ -2,6 +2,7 @@ package se.chalmers.dat265.group1.api;
 
 import se.chalmers.dat265.group1.model.DataModel;
 import se.chalmers.dat265.group1.model.*;
+import se.chalmers.dat265.group1.model.dbEntities.OrderedPart;
 import se.chalmers.dat265.group1.storage.DBInterface;
 import se.chalmers.dat265.group1.storage.PostgresSQLConnector;
 import se.chalmers.dat265.group1.storage.repository.GenericRepository;
