+++ backend/src/main/java/se/chalmers/dat265/group1/api/customer/CustomerAPI.java
@@ -1,6 +1,7 @@
 package se.chalmers.dat265.group1.api.customer;
 
 import se.chalmers.dat265.group1.model.*;
+import se.chalmers.dat265.group1.model.dbEntities.OrderedPart;
 
 import java.util.List;
 
@@ -22,20 +23,4 @@ public interface CustomerAPI {
     Customer updateCustomer(String customerID, Customer customer);
 
     String deleteCustomer(String customerID);
-
-    List<Order> getOrdersFromCustomer(String customerID);
-
-    List<Order> getAllOrders();
-
-    Order getOrder(String orderID);
-
-    Order createNewOrder(Order order);
-
-    List<OrderedPart> getOrderedParts(String orderID);
-
-    Order updateOrder(String orderID, Order order);
-
-    OrderedPart createNewOrderedPart(String orderID, OrderedPart orderedPart);
-
-    OrderedPart updateOrderDetail(String orderID, String orderedPartID, OrderedPart orderedPart);
 }
