+++ backend/src/main/java/se/chalmers/dat265/group1/api/customer/CustomerController.java
@@ -2,6 +2,8 @@ package se.chalmers.dat265.group1.api.customer;
 
 import se.chalmers.dat265.group1.api.ApiController;
 import se.chalmers.dat265.group1.model.*;
+import se.chalmers.dat265.group1.model.dbEntities.OrderedPart;
+
 import java.util.List;
 
 /**
@@ -49,48 +51,5 @@ public class CustomerController extends ApiController implements CustomerAPI {
     public String deleteCustomer(String customerID) {
         customerRepository.deleteObject(Integer.parseInt(customerID));
         return customerID;
-    }
-
-    @Override
-    public List<Order> getOrdersFromCustomer(String customerID) {
-        return orderRepository.getObjects("customerID=" + customerID);
-    }
-
-    @Override
-    public List<Order> getAllOrders() {
-        return orderRepository.getObjects();
-
-    }
-
-    @Override
-    public Order getOrder(String orderID) {
-        return orderRepository.getObject(Integer.parseInt(orderID));
-    }
-
-    @Override
-    public Order createNewOrder(Order order) {
-        return orderRepository.postObject(order);
-    }
-
-    @Override
-    public List<OrderedPart> getOrderedParts(String orderID) {
-        return orderedPartRepository.getObjects("orderID="+orderID);
-    }
-
-    @Override
-    public Order updateOrder(String orderID, Order order) {
-        checkIDs(orderID, order);
-        return orderRepository.updateObject(order);
-    }
-
-    @Override
-    public OrderedPart createNewOrderedPart(String orderID, OrderedPart orderedPart) {
-        return orderedPartRepository.postObject(orderedPart);
-    }
-
-    @Override
-    public OrderedPart updateOrderDetail(String orderID, String orderedPartID, OrderedPart orderedPart) {
-        checkIDs(orderedPartID,orderedPart);
-        return orderedPartRepository.updateObject(orderedPart);
     }
 }
