+++ backend/src/main/java/se/chalmers/dat265/group1/model/Order.java
@@ -1,13 +1,20 @@
 package se.chalmers.dat265.group1.model;
 
 
+import se.chalmers.dat265.group1.model.dbEntities.OrderedPart;
+
 import java.time.Instant;
+import java.util.LinkedList;
+import java.util.List;
 
 public class Order extends DataModel {
 
     private int id;
     private int customerID;
     private String date;
+    private List<OrderedPart> orderedParts = new LinkedList<>();
+
+
 
     public Order() {
         id = -1;
@@ -25,5 +32,9 @@ public class Order extends DataModel {
 
     public String getDate() {
         return date;
+    }
+
+    public List<OrderedPart> getOrderedParts() {
+        return orderedParts;
     }
 }
